Add consistent ring hashing algorithm so Client has a ring of server nodes
