+++ Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "async-memcached"
-version = "0.1.7"
+version = "0.2.0"
 authors = ["Toby Lawrence <toby@nuclearfurnace.com>"]
-edition = "2018"
+edition = "2021"
 readme = "README.md"
 license = "MIT"
 description = "An Tokio-based memcached client for Rust."
@@ -17,6 +17,8 @@ pin-project = "1.0"
 futures = "0.3"
 tokio = { version = "1.26", default-features = false, features = ["io-util"] }
 async-stream = "0.3"
+crc32fast = "1.4.2"
+blake3 = "1.5.1"
 
 [dev-dependencies]
 lazy_static = "1.4"
